refactor: BlackjackGame 패키지 위치 변경 및 사용하지 않는 상수 정리
- BlackjackGame 의 패키지 위치를 service -> domain으로 이동
- 사용하지 않는 상수 BLACKJACK_THRESHOLD 제거
